Refactor dfkde model
